Move identity setup extensions to dedicated namespace
